Move SynchronizingEnumerator to Common\Util
This is a pretty generic utility class that is capable of accepting multiple
IEnumerator<BaseData> instances and synchronizing them into a single
IEnumerator<BaseData> instance. This can be handy for many things. This move is
to enable usage in the QuantConnect.Brokerages projects where the IB history
provider impl needs to synchronize multiple trader bar streams into Slices.
